refacto(provider): move ProviderCapabilities into provider entities
Relocate the ProviderCapabilities value object from the now-empty
_providergateway module into entities.py (as a BaseModel, matching the
other domain types) and delete _providergateway.py. Update importers to
api.domain.provider.entities.
