+++ api/domain/provider/__init__.py
@@ -1,7 +1,6 @@
 from api.domain.provider._provideradapter import ProviderAdapter
 from api.domain.provider._provideradapterbuilder import ProviderAdapterBuilder
 from api.domain.provider._providerclient import ProviderClient, ProviderClientResponse
-from api.domain.provider._providergateway import ProviderCapabilities
 from api.domain.provider._providerloadbalancer import ProviderLoadBalancer
 from api.domain.provider._providermetricslogger import ProviderMetricsLogger
 from api.domain.provider._providerrepository import ProviderRepository
@@ -11,7 +10,6 @@ __all__ = [
     "ProviderAdapterBuilder",
     "ProviderClient",
     "ProviderClientResponse",
-    "ProviderCapabilities",
     "ProviderLoadBalancer",
     "ProviderMetricsLogger",
     "ProviderRepository",
