+++ api/use_cases/provider/_getprovidercapabilities.py
@@ -1,8 +1,8 @@
 from api.domain.embeddings.entities import CreateEmbeddingsBody
 from api.domain.model.entities import ModelType as RouterType
 from api.domain.model.errors import ModelNotFoundError
-from api.domain.provider import ProviderAdapter, ProviderAdapterBuilder, ProviderCapabilities, ProviderClient
-from api.domain.provider.entities import Provider, ProviderOriginalRequest, ProviderOriginalResponse, ProviderType
+from api.domain.provider import ProviderAdapter, ProviderAdapterBuilder, ProviderClient
+from api.domain.provider.entities import Provider, ProviderCapabilities, ProviderOriginalRequest, ProviderOriginalResponse, ProviderType
 from api.domain.provider.errors import ProviderNotReachableError
 from api.utils.variables import EndpointRoute
